+++ src/scripts/train/OLMoE2-large.py
@@ -132,7 +132,7 @@ if __name__ == "__main__":
         train_module_config_builder=build_train_module_config,
         trainer_config_builder=build_trainer_config,
         include_default_evals=False,
-        beaker_image=OLMoCoreBeakerImage.stable_cu126,
+        beaker_image=OLMoCoreBeakerImage.tch271_cu126,
         num_nodes=8,
     )
     main(config_builder=config_builder)